requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.13
djangorestframework==3.13.1
filelock==3.6.0
idna==3.3
importlib-metadata==4.11.3
numpy==1.21.5
platformdirs==2.5.1
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.3
zipp==3.7.0
